Move BotoHelper up so it can be used elsewhere
